BUGFIX - delay slot status could be set by squashed instructions
To lessen the cost of flushing the pipeline, each stage is paired with
a valid bit. This bit must be consulted if and only the stage result
is used to guide pipeline control (flushes and restarts) or commit
results.

The delay slot summary is use to calculate the restart address and
thus must take the valid bit into account.