Co-simulation take IO-space values from the RTL model.
Keeping the IO model up to date in the ISA simulator is unsuitainable
and waste of time. While we still need minimal IO in the ISA simulator
to run test apps, timing between the two will not match.

The only flip side of this change is to lessen the amount of testing
the RTL IO devices gets from cosimulation.