CORE: lower the perf counter penalty and synthesize shift ops
Stop using Altera specific primitives for shifts in preparation of
streching them over two cycles.

Get the perf counters out of the critical path. It's completely
irrelevant whether the perf counters are phase shifted, so register
the occurance of the event, giving us a full cycle to just increment
the counter. So far it has only been done for perf_branch_hazards, but
the others should get a similar treatment (and in fact the counters
themselves should migrate to a dedicated module).