TARGET DE2-70: add a video interface, 1024x768x8 @ 60 Hz
The palette is currently fixed at RGB332, the mode is hardcoded, and
the framebuffer location is fixed at 0x4010000.

Note, the video clock isn't synchronized to the system clock and we
use a FIFO to transfer pixel data between the two domains and a
simple double FF to reset the fetch for every frame.