YARI bugfix: the shadow of a ld/st hazard wasn't flushed
Loads from IO space (and stores to a busy peripheral) have variable
latency and thus the pipeline must block until the resource is ready.
In YARI this is done by restarting the whole pipeline.  Unfortunately,
the instruction in the EX stage wasn't flushed until a cycle later
which was fatal when that happend to be a any store or an IO load.

Pipelining is trivial in the absense of hazards...